Use `ffi.errno`
Take a look at https://luajit.org/ext_ffi_api.html (section `Utility Functions`) for what `ffi.errno` does :3

Reason for change:
IEEE standard does not provide guarantee that `errno` is an actual linkable symbol

From https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/errno.h.html

The symbol errno shall expand to a modifiable lvalue of type int. **It is unspecified whether errno is a macro or an identifier declared with external linkage.** If a macro definition is suppressed in order to access an actual object, or a program defines an identifier with the name errno, the behavior is undefined.